updated mobile navbar, profile and changed navmenu
