+++ frontend-next-migration/src/shared/ui/NavMenu/index.ts
@@ -1,3 +1,5 @@
 export { default as NavMenu } from './ui/NavMenu';
-export type { NavMenuProps, DropdownItem } from './ui/NavMenu';
+export type { NavMenuProps } from './ui/NavMenu';
+export type { INavMenuDropdown, INavMenuLink, INavMenuElement, INavMenuItem } from './model/types';
+export { NavMenuItemType } from './model/types';
 export type { DropDownElementASTextOrLink } from '@/shared/ui/DropdownWrapperV2';
