+++ frontend-next-migration/src/shared/ui/NavMenu/ui/NavMenu.module.scss
@@ -78,3 +78,10 @@
     color: var(--primary-color);
   }
 }
+.active {
+  color: var(--primary-color) !important;
+}
+.activeDropdown {
+  color: var(--primary-color) !important;
+  text-decoration: underline !important;
+}
